requirements.txt
django==3.2
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.13.1
drf-yasg==1.21.3
psycopg2-binary==2.9.3
requests==2.28.1